[refactoring] validator class 분리 및 주석 보강
